requirements.txt
requests
Jinja2==2.10.1
future==0.17.1
watchdog==0.9.0
PyGithub==1.43.6
GitPython==2.1.11
firebase-admin==2.16.0
google-cloud-firestore==0.32.1
cryptography==2.6.1
pytest==4.4.1
